requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic>=2.5.0
pydantic-settings

# RAG Components
chromadb
sentence-transformers
groq

# Google Sheets
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2

# Utilities
python-dotenv
pandas
numpy
openpyxl

# Background Tasks
APScheduler

# Production
slowapi
python-multipart
redis>=5.0.0
better-profanity
gunicorn  # Multi-worker production server for 200+ users
